module側で、@Singletonを指定するように変更しました。
