+++ app/src/main/java/com/example/gourmetsearchercompose/di/location/SearchLocationModule.kt
@@ -1,6 +1,8 @@
 package com.example.gourmetsearchercompose.di.location
 
 import android.content.Context
+import com.example.gourmetsearchercompose.repository.SearchLocationRepository
+import com.example.gourmetsearchercompose.repository.SearchLocationRepositoryImpl
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
 import dagger.Module
@@ -8,18 +10,31 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
 
 /** 位置情報のモジュール */
 @Module
 @InstallIn(SingletonComponent::class)
-object SearchLocationServicesModule {
+object SearchLocationModule {
     /**
-     * 位置情報のリポジトリを提供する
+     * 位置情報のプロバイダーを提供する
      * @param context コンテキスト
-     * @return 位置情報のリポジトリ
+     * @return FusedLocationProviderClient
      */
     @Provides
+    @Singleton
     fun provideFusedLocationProviderClient(
         @ApplicationContext context: Context,
     ): FusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(context)
+
+    /**
+     * 位置情報のリポジトリを提供する
+     * @param locationProvider 位置情報のプロバイダー
+     * @return 位置情報のリポジトリ
+     */
+    @Provides
+    @Singleton
+    fun provideLocationRepository(
+        locationProvider: FusedLocationProviderClient
+    ): SearchLocationRepository = SearchLocationRepositoryImpl(locationProvider)
 }
