+++ app/src/main/java/com/example/gourmetsearchercompose/di/network/NetWorkModule.kt
@@ -2,6 +2,7 @@ package com.example.gourmetsearchercompose.di.network
 
 import android.content.Context
 import com.example.gourmetsearchercompose.R
+import com.example.gourmetsearchercompose.service.HotPepperGourmetApiService
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import dagger.Module
@@ -11,6 +12,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
+import javax.inject.Singleton
 
 /** NetWorkのモジュール */
 @Module
@@ -21,6 +23,7 @@ object NetWorkModule {
      * @return Moshi
      */
     @Provides
+    @Singleton
     fun provideMoshi(): Moshi =
         Moshi
             .Builder()
@@ -34,6 +37,7 @@ object NetWorkModule {
      * @return Retrofit
      */
     @Provides
+    @Singleton
     fun provideRetrofit(
         @ApplicationContext context: Context,
         moshi: Moshi,
@@ -43,4 +47,15 @@ object NetWorkModule {
             .baseUrl(context.getString(R.string.restaurant_list_hot_pepper_url))
             .addConverterFactory(MoshiConverterFactory.create(moshi))
             .build()
+
+    /**
+     * HotPepperGourmetApiServiceを提供
+     * @param retrofit Retrofit
+     * @return HotPepperGourmetApiService
+     */
+    @Provides
+    @Singleton // Singletonスコープを明示
+    fun provideHotPepperService(
+        retrofit: Retrofit
+    ): HotPepperGourmetApiService = retrofit.create(HotPepperGourmetApiService::class.java)
 }
