+++ gradle/libs.versions.toml
@@ -1,6 +1,6 @@
 [versions]
 accompanistPermissions = "0.36.0"
-agp = "8.7.0-beta01"
+agp = "8.7.2"
 coil = "2.7.0"
 coreTesting = "2.2.0"
 datastorePreferences = "1.1.1"
